Create bundle dir if it doesn't exist and clean it up at the end.
- TODO: find a way to make bundleDir more random. Can we use information
from the groot create output? Right now, we are running MkdirAll which
will not fail if the bundle dir exists, but we would want these
directories to be different for each container since the bundle dir name
is generated from the container id which is generated from the current
time.